+++ R/mark_intensity.R
@@ -3,8 +3,9 @@
 #' This function adds several columns to a dataset that contains accelerometer counts data. 
 #'     These columns concern respectively sedentary time (SED), light physical activity time (LPA), 
 #'     moderate physical activity time (MPA), vigorous physical activity time (VPA), 
-#'     metabolic equivalent of task (METs), kilocalories (kcal), and MET-hours when 
-#'     time is spent in moderate-to-vigorous physical activity. For the SED, LPA, MPA, 
+#'     metabolic equivalent of task (METs), kilocalories (kcal), MET-hours when 
+#'     time is spent in moderate-to-vigorous physical activity, and step-based metrics (maximum step accumulation  
+#'     and peak step accumulation metrics). For the SED, LPA, MPA, 
 #'     and VPA columns, the function provides, for each epoch, the numeric value 1 when 
 #'     the value of the configured counts variable corresponds to the criteria of the 
 #'     considered category (e.g., <150 counts/min for sedentary behavior); if not, 0 
@@ -17,12 +18,12 @@
 #'     that uses sex, age, and weight inputs, and one of the equations retrieved from the paper by Henry et al. (2005; doi: 10.1079/PHN2005801).
 #'     MET-hours are obtained by multiplying METs by time related to each epoch (here 1/60e of an hour), only when the MET value is  >=3.
 #'
-#' @param data A dataframe obtained using the \code{\link{prepare_dataset}} function.
+#' @param data A dataframe obtained using the \code{\link{prepare_dataset}} and then the \code{\link{mark_wear_time}} functions.
 #' @param col_axis A character value to indicate the name of the variable to be used for determining intensity categories. 
 #' @param equation A character string to indicate the equation to be used for estimating METs.
 #' @param sed_cutpoint A numeric value below which time is considered as spent in sedentary behavior.
-#' @param mpa_cutpoint A numeric value above which time is considered as spent in moderate physical activity.
-#' @param vpa_cutpoint A numeric value above which time is considered as spent in vigorous physical activity.
+#' @param mpa_cutpoint A numeric value at and above which time is considered as spent in moderate physical activity.
+#' @param vpa_cutpoint A numeric value at and above which time is considered as spent in vigorous physical activity.
 #' @param age A numeric value in yr.
 #' @param weight A numeric value in kg.
 #' @param sex A character value.
@@ -32,8 +33,9 @@
 #' @examples
 #' file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
 #' mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
+#' mydata_with_wear_marks <- mark_wear_time(dataset = mydata)
 #' mark_intensity(
-#'     data = mydata, 
+#'     data = mydata_with_wear_marks, 
 #'     col_axis = "vm", 
 #'     equation = "Sasaki et al. (2011) [Adults]",
 #'     sed_cutpoint = 200, 
@@ -41,7 +43,8 @@
 #'     vpa_cutpoint = 6167, 
 #'     age = 32,
 #'     weight = 67,
-#'     sex = "male")
+#'     sex = "male",
+#'     col_steps = "steps")
 #' 
 mark_intensity <- function(data, 
                            col_axis = "vm", 
@@ -54,7 +57,8 @@ mark_intensity <- function(data,
                                         "Santos-Lozano et al. (2013) [Older adults]"),
                            age = 40, 
                            weight = 70, 
-                           sex = c("male", "female", "undefined")) {
+                           sex = c("male", "female", "undefined"),
+                           col_steps = "steps") {
   
   equation <- match.arg(equation)
   sex <- match.arg(sex)
@@ -66,16 +70,39 @@ mark_intensity <- function(data,
     df <-
       data %>%
       dplyr::mutate(
-      SED = ifelse(.data[[col_axis]] < sed_cutpoint, 1, 0),
-      LPA = ifelse(.data[[col_axis]] >= sed_cutpoint & .data[[col_axis]] < mpa_cutpoint, 1, 0),
-      MPA = ifelse(.data[[col_axis]] >= mpa_cutpoint & .data[[col_axis]] < vpa_cutpoint, 1, 0), 
-      VPA = ifelse(.data[[col_axis]] >= vpa_cutpoint, 1, 0),
-      METS = compute_mets(data = .data, equation = equation, weight = weight, sex = sex),
-      kcal = ifelse(SED == 0, METS * bmr_kcal_min, bmr_kcal_min),
+        SED = ifelse(.data[[col_axis]] < sed_cutpoint, 1, 0),
+        LPA = ifelse(.data[[col_axis]] >= sed_cutpoint & .data[[col_axis]] < mpa_cutpoint, 1, 0),
+        MPA = ifelse(.data[[col_axis]] >= mpa_cutpoint & .data[[col_axis]] < vpa_cutpoint, 1, 0), 
+        VPA = ifelse(.data[[col_axis]] >= vpa_cutpoint, 1, 0),
+        METS = compute_mets(data = .data, equation = equation, weight = weight, sex = sex),
+        kcal = ifelse(SED == 0, METS * bmr_kcal_min, bmr_kcal_min),
       
-  # Computing MET-hr corresponding to MVPA only, for each epoch
-    mets_hours_mvpa = ifelse(METS >=3, 1/60 * METS, 0))
+      # Computing MET-hr corresponding to MVPA only, for each epoch
+        mets_hours_mvpa = ifelse(METS >= 3, 1/60 * METS, 0),
+      
+      # Applying moving averages for future computations of step-based metrics
+        accum_steps_60min = zoo::rollmean(data[[col_steps]], align = "center", k = 60L, fill = NA),
+        accum_steps_30min = zoo::rollmean(data[[col_steps]], align = "center", k = 30L, fill = NA),
+        accum_steps_20min = zoo::rollmean(data[[col_steps]], align = "center", k = 20L, fill = NA),
+        accum_steps_5min  = zoo::rollmean(data[[col_steps]], align = "center", k = 5L,  fill = NA),
+        accum_steps_1min  = zoo::rollmean(data[[col_steps]], align = "center", k = 1L,  fill = NA)
+  )
     
+  # Marking the bouts based on intensity categories
+    df$intensity_category <- ifelse(df$SED == 1, "SED", ifelse(df$LPA == 1, "LPA", ifelse(df$MPA == 1 | df$VPA == 1, "MVPA", "none")))
+    df$bout <- vector("double", length = nrow(df))
+    df$bout[1] <- 1
+    for (i in 2:(nrow(df) - 1)) {
+      if (df$intensity_category[i] != df$intensity_category[i-1]) {
+        df$bout[i] <- df$bout[i-1] + 1
+      } else {
+        df$bout[i] <- df$bout[i-1]
+      }
+    }
+    df$bout[nrow(df)] <- df$bout[nrow(df)-1] 
+    
+    
+  # Providing information about the parameters used for computing results
     message(paste0("You have computed intensity metrics using the following inputs: 
     axis = ", col_axis, "
     sed_cutpoint = ", sed_cutpoint, "
@@ -85,7 +112,8 @@ mark_intensity <- function(data,
     age = ", age, "
     weight = ", weight, "
     sex = ", sex))
-  
+   
+  # Returning dataframe 
     return(df)
     
 }