+++ NEWS.md
@@ -1,5 +1,7 @@
 # activAnalyzer (development version)
 * Add a block of code to stop the app when the last user leaves the app.
+* Modify the plot for visualizing data with wear time so that each day of measurement be on one line.
+* Add step-based metrics to the results.
 
 # activAnalyzer 0.0.3
 * Add alert message when closing the web window.